Patch to fix control key handling in Firefox
SudokuPad assumes that the event.ctrlKey will be true on a Key Down of
ControlLeft or ControlRight, ie that "the control key was pressed at the
time the key down event was generated".  But in Firefox on Linux this
is not always true; some Key Down ControlLeft / ControlRight events are
generated with event.ctrlKey = true, and some with event.ctrlKey = false.

Similarly SudokuPad assumes that event.ctrlKey will be false on a
Key Up of ControlLeft or ControlRight, ie that hte "the control key
was not pressed at the time the key up event was generated".  But in
Firefox on Linux this is not always true; some Key Up ControlLeft /
ControlRight events are generated with event.ctrlKey = false, and some
with event.ctrlKey = true.

This patch reduces the impacts of the false assumptions about key events
being made by SudokuPad.