Expand tinyurl.com and app.crackingthecryptic.com
Older GAS Leaks (up to about November 2023) used tinyurl.com extensively
instead of (and/or mixed in with) the SudokuPad shortener.  And a bunch
of puzzles in 2023 used app.crackingthecryptic.com instead of
sudokupad.app.

Also try harder to reject invalid references to sites which are not actually
puzzle URLs.