Extract URLs from Sudoku Clover Puzzle Packs
And expand to full URLs.  Similar to gasfullurls, but Sudoku Clover
Patreon Puzzle Packs are formatted differently.

(The earlier Puzzle Packs have the full solving links in the PDFs, but
the later ones -- since about 2024 -- do not.  So for consistency all
the full Puzzle URLs are extracted by expanding the TinyURLs.)