Extract Puzzle URLs from Sudoku Theory PDFs
The Sudoku Theory PDFs are formatted with the name of the solving package
as the "link text", and the actual solving link hidden behind that
link text.  This means we cannot use pdftotext to extract the bare
short links, but instead need to use pdftohtml to extract a full
version with <a href="...">...</a> links in it, and then parse the
HTML.  The parser makes a lot of assumptions, but mostly works for
the Sudoku Theory (Philip Newman) puzzles encountered so far.